Migrate OID4VCJWTIssuerEndpointDisabledTest and OID4VCSdJwtIssuingEndpointDisabledTest
Closes #48743
