+++ tests/base/src/test/java/org/keycloak/tests/oid4vc/issuance/signing/OID4VCJWTIssuerEndpointDisabledTest.java
@@ -1,4 +1,4 @@
-package org.keycloak.testsuite.oid4vc.issuance.signing;
+package org.keycloak.tests.oid4vc.issuance.signing;
 
 import java.util.function.Consumer;
 
@@ -6,36 +6,64 @@ import jakarta.ws.rs.core.Response;
 
 import org.keycloak.protocol.oid4vc.issuance.OID4VCIssuerEndpoint;
 import org.keycloak.protocol.oid4vc.model.CredentialRequest;
-import org.keycloak.representations.idm.ClientRepresentation;
-import org.keycloak.representations.idm.RealmRepresentation;
 import org.keycloak.services.CorsErrorResponseException;
 import org.keycloak.services.managers.AppAuthManager;
+import org.keycloak.testframework.annotations.KeycloakIntegrationTest;
+import org.keycloak.testframework.remote.runonserver.InjectRunOnServer;
+import org.keycloak.testframework.remote.runonserver.RunOnServerClient;
+import org.keycloak.tests.oid4vc.OID4VCIssuerEndpointTest;
+import org.keycloak.tests.oid4vc.OID4VCIssuerTestBase;
+import org.keycloak.testsuite.util.AccountHelper;
 import org.keycloak.util.JsonSerialization;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
- * Tests for OID4VCIssuerEndpoint with OID4VCI disabled for SD-JWT format
+ * Tests for OID4VCIssuerEndpoint with OID4VCI disabled.
  */
-public class OID4VCSdJwtIssuingEndpointDisabledTest extends OID4VCIssuerEndpointTest {
+@KeycloakIntegrationTest(config = OID4VCIssuerTestBase.VCTestServerConfig.class)
+public class OID4VCJWTIssuerEndpointDisabledTest extends OID4VCIssuerEndpointTest {
 
-    @Override
-    protected boolean shouldEnableOid4vci(RealmRepresentation testRealm) {
-        return false;
+    @InjectRunOnServer
+    RunOnServerClient runOnServer;
+
+    @BeforeEach
+    void disableVerifiedCredentialsOnTestRealm() {
+        testRealm.updateWithCleanup(realm -> realm.verifiableCredentialsEnabled(false));
     }
 
-    @Override
-    protected boolean shouldEnableOid4vci(ClientRepresentation testClient) {
-        return false;
+
+    @AfterEach
+    void logout() {
+        AccountHelper.logout(testRealm.admin(), "john");
+    }
+
+    /**
+     * When verifiable credentials are disabled at the realm level, OID4VCI endpoints
+     * must reject calls regardless of the client configuration.
+     */
+    @Test
+    public void testRealmDisabledEndpoints() {
+        testWithBearerToken(token -> runOnServer.run((session) -> {
+            AppAuthManager.BearerTokenAuthenticator authenticator = new AppAuthManager.BearerTokenAuthenticator(session);
+            authenticator.setTokenString(token);
+            OID4VCIssuerEndpoint issuerEndpoint = prepareIssuerEndpoint(session, authenticator);
+
+            // Nonce endpoint should be forbidden when OID4VCI is disabled for the realm
+            CorsErrorResponseException nonceException = Assertions.assertThrows(CorsErrorResponseException.class, issuerEndpoint::getCNonce);
+            assertEquals(Response.Status.FORBIDDEN.getStatusCode(), nonceException.getResponse().getStatus(), "Realm-disabled OID4VCI should return 403 for nonce endpoint");
+        }));
     }
 
     @Test
     public void testClientNotEnabled() {
-        testWithBearerToken(token -> testingClient.server(TEST_REALM_NAME).run((session) -> {
+        testWithBearerToken(token -> runOnServer.run((session) -> {
             AppAuthManager.BearerTokenAuthenticator authenticator = new AppAuthManager.BearerTokenAuthenticator(session);
             authenticator.setTokenString(token);
             OID4VCIssuerEndpoint issuerEndpoint = prepareIssuerEndpoint(session, authenticator);
@@ -46,8 +74,9 @@ public class OID4VCSdJwtIssuingEndpointDisabledTest extends OID4VCIssuerEndpoint
             );
             assertEquals(Response.Status.FORBIDDEN.getStatusCode(), offerUriException.getResponse().getStatus(), "Should fail with 403 Forbidden when client is not OID4VCI-enabled");
 
+            // Test requestCredential
             CredentialRequest credentialRequest = new CredentialRequest()
-                    .setCredentialConfigurationId(sdJwtTypeCredentialConfigurationIdName);
+                    .setCredentialIdentifier(jwtTypeCredentialScopeName);
             String requestPayload;
             try {
                 requestPayload = JsonSerialization.writeValueAsString(credentialRequest);
